requirements.txt
selenium>=3.1
requests>=2.21
pytest>=4.1.0
allure-pytest>=2.5
